backend/requirements.txt
flask==2.0.1
werkzeug==2.0.1
flask-cors==3.0.10
gunicorn==23.0.0
python-dotenv==1.0.0
# LangChain packages are now installed directly in Dockerfile
# langchain>=0.1.0
# langchain-community>=0.1.0
# langchain-aws>=0.1.0
requests>=2.27.1
pydantic>=1.9.0
numpy>=1.24.0
pandas>=1.5.3
tqdm>=4.62.3
PyPDF2==3.0.1
docx2txt==0.8
openpyxl==3.1.2